backend/requirements.txt
# Kafka dependencies
kafka-python==2.0.2

# OCR dependencies
pytesseract==0.3.10
Pillow==10.0.1

# Image processing and ML
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3

# FAISS for similarity search
faiss-cpu==1.7.4

# FastAPI for services
fastapi==0.104.1
uvicorn==0.24.0

# Flink dependencies
apache-flink==1.17.1
pyflink==1.17.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Optional: For production deployment
gunicorn==21.2.0 